[auto-materialize][refactor] Move "asset reconciliation" logic into an AssetDaemonContext object (#15427)
## Summary & Motivation

This moves all of the asset reconciliation logic into a more manageable
AssetDaemonContext -- basically a pure copy/paste of the methods, but
with the added benefit of removing a bunch of the arguments to some of
these functions, as they can just become properties of the class.

This is just making room for a more fundamental change to the
reconciliation logic in the next PR.

## How I Tested These Changes
