+++ python_modules/dagster/dagster/_core/execution/asset_backfill.py
@@ -19,9 +19,9 @@ from typing import (
 )
 
 from dagster import _check as check
+from dagster._core.definitions.asset_daemon_context import build_run_requests
 from dagster._core.definitions.asset_graph import AssetGraph
 from dagster._core.definitions.asset_graph_subset import AssetGraphSubset
-from dagster._core.definitions.asset_reconciliation_sensor import build_run_requests
 from dagster._core.definitions.asset_selection import AssetSelection
 from dagster._core.definitions.assets_job import is_base_asset_job_name
 from dagster._core.definitions.events import AssetKey, AssetKeyPartitionKey
