+++ python_modules/dagster/dagster_tests/definitions_tests/asset_reconciliation_tests/test_asset_reconciliation_sensor.py
@@ -12,11 +12,11 @@ from dagster import (
     repository,
 )
 from dagster._check import CheckError
+from dagster._core.definitions.asset_daemon_context import build_auto_materialize_asset_evaluations
 from dagster._core.definitions.asset_graph import AssetGraph
 from dagster._core.definitions.asset_reconciliation_sensor import (
     AutoMaterializeAssetEvaluation,
     build_asset_reconciliation_sensor,
-    build_auto_materialize_asset_evaluations,
 )
 from dagster._core.definitions.events import AssetKey, AssetKeyPartitionKey
 from dagster._core.definitions.time_window_partitions import (
